Extract decision module
